Some versions of scipy don't like 'float' dtype ro sparse rand.

+++ sklearn/manifold/tests/test_isomap.py
@@ -126,8 +126,7 @@ def test_isomap_clone_bug():
                      n_neighbors)
 
 def test_sparse_input():
-    X = sparse_rand(100, 3, density=0.1, format='csr',
-                    dtype=np.float)
+    X = sparse_rand(100, 3, density=0.1, format='csr')
 
     # Should not error
     for eigen_solver in eigen_solvers:
